+++ src/main/java/lotto/service/calculator/BonusCalculator.java
@@ -1,35 +1,23 @@
-package lotto.service.generator;
+package lotto.service.calculator;
 
 import java.util.ArrayList;
 import java.util.List;
 import lotto.domain.Bonus;
 import lotto.domain.Lotto;
 
-public class ResultGenerator {
+public class BonusCalculator {
 
     private static final Integer TRUE = 1;
     private static final Integer FALSE = 0;
 
-    private final List<Integer> winningResult = new ArrayList<>();
     private final List<Integer> bonusResult = new ArrayList<>();
 
-    private ResultGenerator(List<Lotto> lottoTicket, Lotto winning, Bonus bonus) {
-        compareTicketToWinning(lottoTicket, winning);
+    private BonusCalculator(List<Lotto> lottoTicket, Bonus bonus) {
         compareTicketToBonus(lottoTicket, bonus);
     }
 
-    public static ResultGenerator create(List<Lotto> lottoTicket, Lotto winning, Bonus bonus) {
-        return new ResultGenerator(lottoTicket, winning, bonus);
-    }
-
-    private void compareTicketToWinning(List<Lotto> lottoTicket, Lotto winning) {
-        lottoTicket.forEach(lotto -> countWinning(lotto, winning));
-    }
-
-    private void countWinning(Lotto lotto, Lotto winning) {
-        winningResult.add((int) lotto.getNumbers().stream()
-                .filter(number -> winning.getNumbers().contains(number))
-                .count());
+    public static BonusCalculator create(List<Lotto> lottoTicket, Bonus bonus) {
+        return new BonusCalculator(lottoTicket, bonus);
     }
 
     private void compareTicketToBonus(List<Lotto> lottoTicket, Bonus bonus) {
@@ -46,10 +34,6 @@ public class ResultGenerator {
 
     private boolean IsContainBonus(List<Integer> lotto, Integer bonus) {
         return lotto.contains(bonus);
-    }
-
-    public List<Integer> getWinningResult() {
-        return winningResult;
     }
 
     public List<Integer> getBonusResult() {
